requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
pymupdf
numpy
pydantic
requests
python-multipart
streamlit
typing-extensions